🫃A TON of fixes

+++ cogs/admin/say.py
@@ -12,9 +12,11 @@ class SayCommand(commands.Cog):
 	@app_commands.default_permissions(manage_messages=True)
 
 	async def say(self, ctx, *, text: str):
-		await ctx.send("_ _", ephemeral=True, delete_after=0)
 		await ctx.channel.send(text)
-		await ctx.message.delete()
+		if ctx.interaction:
+			await ctx.send("_ _", ephemeral=True, delete_after=0)
+		else:
+			await ctx.message.delete()
 
 	@say.error
 	async def say_error(self, ctx, error):
